Delete team_test file and rename teams_test to team_test
